Revert "[Asset graph] Add tree view (#31929)"
This reverts commit 111d790778b65e54c6c0b2cb88fd9e0ca386af19.
