+++ js_modules/dagster-ui/packages/ui-core/src/asset-graph/sidebar/AssetSidebarNode.tsx
@@ -1,13 +1,5 @@
-import {
-  Box,
-  Colors,
-  Icon,
-  MiddleTruncate,
-  UnstyledButton,
-  useDelayedState,
-} from '@dagster-io/ui-components';
+import {Box, Colors, Icon, MiddleTruncate, UnstyledButton} from '@dagster-io/ui-components';
 import * as React from 'react';
-import {useMemo} from 'react';
 import {observeEnabled} from 'shared/app/observeEnabled.oss';
 import styled from 'styled-components';
 
@@ -28,7 +20,6 @@ import {GraphData, GraphNode} from '../Utils';
 
 type AssetSidebarNodeProps = {
   fullAssetGraphData?: GraphData;
-  graphData: GraphData;
   node: GraphNode | FolderNodeNonAssetType;
   level: number;
   toggleOpen: () => void;
@@ -40,40 +31,17 @@ type AssetSidebarNodeProps = {
   explorerPath: ExplorerPath;
   onChangeExplorerPath: (path: ExplorerPath, mode: 'replace' | 'push') => void;
   onFilterToGroup: (group: AssetGroup) => void;
-  viewType: 'tree' | 'group';
-  direction?: 'root-to-leaf' | 'leaf-to-root';
-  collapseAllNodes?: () => void;
 };
 
 export const AssetSidebarNode = (props: AssetSidebarNodeProps) => {
-  const {
-    node,
-    level,
-    toggleOpen,
-    isOpen,
-    selectThisNode,
-    viewType,
-    graphData,
-    direction = 'root-to-leaf',
-    collapseAllNodes,
-  } = props;
+  const {node, level, toggleOpen, isOpen, selectThisNode} = props;
   const isGroupNode = 'groupNode' in node;
   const isLocationNode = 'locationName' in node;
   const isAssetNode = !isGroupNode && !isLocationNode;
 
-  const connectedNodes =
-    direction === 'root-to-leaf'
-      ? Object.keys(graphData.downstream[node.id] ?? {})
-      : Object.keys(graphData.upstream[node.id] ?? {});
   const elementRef = React.useRef<HTMLDivElement | null>(null);
 
-  const showArrow = useMemo(() => {
-    if (viewType === 'tree') {
-      // If the connected node is not in graphData then it is filtered out by the asset selection
-      return connectedNodes.filter((id) => graphData.nodes[id]).length > 0;
-    }
-    return !isAssetNode && !('openAlways' in node && node.openAlways);
-  }, [isAssetNode, node, viewType, connectedNodes, graphData.nodes]);
+  const showArrow = !isAssetNode && !('openAlways' in node && node.openAlways);
 
   return (
     <Box ref={elementRef} padding={{left: 8, right: 12}}>
@@ -114,7 +82,7 @@ export const AssetSidebarNode = (props: AssetSidebarNodeProps) => {
             <div style={{width: 18}} />
           )}
           {isAssetNode ? (
-            <AssetSidebarAssetLabel {...props} node={node} collapseAllNodes={collapseAllNodes} />
+            <AssetSidebarAssetLabel {...props} node={node} />
           ) : isGroupNode ? (
             <AssetSidebarGroupLabel {...props} node={node} />
           ) : (
@@ -134,7 +102,6 @@ type AssetSidebarAssetLabelProps = {
   isSelected: boolean;
   explorerPath: ExplorerPath;
   onChangeExplorerPath: (path: ExplorerPath, mode: 'replace' | 'push') => void;
-  collapseAllNodes?: () => void;
 };
 
 const AssetSidebarAssetLabel = ({
@@ -145,7 +112,6 @@ const AssetSidebarAssetLabel = ({
   selectNode,
   explorerPath,
   onChangeExplorerPath,
-  collapseAllNodes,
 }: AssetSidebarAssetLabelProps) => {
   const {menu, dialog} = useAssetNodeMenu({
     graphData: fullAssetGraphData,
@@ -153,12 +119,7 @@ const AssetSidebarAssetLabel = ({
     selectNode,
     explorerPath,
     onChangeExplorerPath,
-    collapseAllNodes,
   });
-
-  // This is a hack: prevent showing the popover for a second in case we are scrolling very fast.
-  // This is to workaround a bug with blueprint Popover's ResizeSensor causing an infinite loop during the animated scrolling.
-  const canShowPopover = useDelayedState(1000);
 
   return (
     <ContextMenuWrapper stopPropagation menu={menu} wrapperOuterStyles={{width: '100%'}}>
@@ -168,11 +129,7 @@ const AssetSidebarAssetLabel = ({
         icon={
           observeEnabled() ? (
             <div style={{marginLeft: -8, marginRight: -8}}>
-              <AssetHealthSummary
-                iconOnly
-                assetKey={node.assetKey}
-                canShowPopover={canShowPopover}
-              />
+              <AssetHealthSummary iconOnly assetKey={node.assetKey} />
             </div>
           ) : (
             <StatusDot node={node} />
@@ -183,8 +140,7 @@ const AssetSidebarAssetLabel = ({
       <ExpandMore onClick={triggerContextMenu}>
         <Icon name="more_horiz" color={Colors.accentGray()} />
       </ExpandMore>
-      {/* Stop propagation of clicks from the dialog otherwise they will trigger the click handler from `ItemContainer` */}
-      <div onClick={(e) => e.stopPropagation()}>{dialog}</div>
+      {dialog}
     </ContextMenuWrapper>
   );
 };
@@ -272,7 +228,7 @@ const FocusableLabelContainer = ({
 const BoxWrapper = ({level, children}: {level: number; children: React.ReactNode}) => {
   const wrapper = React.useMemo(() => {
     let sofar = children;
-    for (let i = 1; i < level; i++) {
+    for (let i = 0; i < level; i++) {
       sofar = (
         <Box
           padding={{left: 8}}
