+++ js_modules/dagster-ui/packages/ui-core/src/assets/AssetHealthSummary.tsx
@@ -40,39 +40,17 @@ import {TimeFromNow} from '../ui/TimeFromNow';
 import {numberFormatter} from '../ui/formatters';
 
 export const AssetHealthSummary = React.memo(
-  ({
-    assetKey,
-    iconOnly,
-    canShowPopover = true,
-  }: {
-    assetKey: {path: string[]};
-    iconOnly?: boolean;
-    canShowPopover?: boolean;
-  }) => {
+  ({assetKey, iconOnly}: {assetKey: {path: string[]}; iconOnly?: boolean}) => {
     if (!observeEnabled()) {
       return null;
     }
 
-    return (
-      <AssetHealthSummaryImpl
-        assetKey={assetKey}
-        iconOnly={iconOnly}
-        canShowPopover={canShowPopover}
-      />
-    );
+    return <AssetHealthSummaryImpl assetKey={assetKey} iconOnly={iconOnly} />;
   },
 );
 
 const AssetHealthSummaryImpl = React.memo(
-  ({
-    assetKey,
-    iconOnly,
-    canShowPopover,
-  }: {
-    assetKey: {path: string[]};
-    iconOnly?: boolean;
-    canShowPopover?: boolean;
-  }) => {
+  ({assetKey, iconOnly}: {assetKey: {path: string[]}; iconOnly?: boolean}) => {
     const {liveData} = useAssetHealthData(assetKey);
     const health = liveData?.assetHealth;
 
@@ -107,11 +85,7 @@ const AssetHealthSummaryImpl = React.memo(
     }
 
     return (
-      <AssetHealthSummaryPopover
-        assetKey={assetKey}
-        health={health}
-        canShowPopover={canShowPopover}
-      >
+      <AssetHealthSummaryPopover assetKey={assetKey} health={health}>
         {content()}
       </AssetHealthSummaryPopover>
     );
@@ -122,12 +96,10 @@ export const AssetHealthSummaryPopover = ({
   health,
   assetKey,
   children,
-  canShowPopover = true,
 }: {
   health: AssetHealthFragment['assetHealth'] | undefined;
   assetKey: AssetKeyInput;
   children: React.ReactNode;
-  canShowPopover?: boolean;
 }) => {
   const {iconName, iconColor, text} = useMemo(() => {
     return statusToIconAndColor[health?.assetHealth ?? 'undefined'];
@@ -167,7 +139,6 @@ export const AssetHealthSummaryPopover = ({
   return (
     <Popover
       interactionKind="hover"
-      isOpen={canShowPopover ? undefined : false}
       content={
         <div onClick={(e) => e.stopPropagation()}>
           <Box padding={12} flex={{direction: 'row', alignItems: 'center', gap: 6}} border="bottom">
